Use testinit instead
