--- kubinit/testing-server/.dockerignore
@@ -1,42 +1,0 @@
-# Git
-.git
-.gitignore
-
-# Documentation
-*.md
-!README.md
-
-# Python
-__pycache__
-*.pyc
-*.pyo
-*.pyd
-.Python
-env
-venv
-.venv
-pip-log.txt
-pip-delete-this-directory.txt
-
-# IDEs
-.vscode
-.idea
-*.swp
-*.swo
-*~
-
-# OS
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.db
-
-# Logs
-*.log
-
-# Temporary files
-*.tmp
-*.temp
