+++ testinit/docker-compose.yml
@@ -1,0 +1,37 @@
+services:
+  provisioning-agent:
+    build:
+      context: ../
+      dockerfile: ./testinit/Dockerfile
+    privileged: true
+    cgroup: host
+    tmpfs:
+      - /run
+      - /tmp
+    volumes:
+      - /sys/fs/cgroup:/sys/fs/cgroup:rw
+      - /run/dbus/system_bus_socket:/run/dbus/system_bus_socket:rw
+    stdin_open: true
+    tty: true
+    container_name: azureinit-provisioning-agent
+    cap_add:
+      - SYS_ADMIN
+    environment:
+      - PATH_HOSTNAMECTL=/usr/bin/hostnamectl
+    networks:
+      - agent-network
+      - wireserver-network
+    logging:
+      driver: "json-file"
+      options:
+        max-size: "10m"
+        max-file: "3"
+
+networks:
+  agent-network:
+    external: true
+    name: imds-network
+
+  wireserver-network:
+    external: true
+    name: wireserver-network
