+++ testinit/start-all.sh
