+++ testinit/stop-all.sh
