+++ testinit/testing-server/docker-compose.yml
@@ -1,0 +1,47 @@
+services:
+  testing-server:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    container_name: azure-testing-server
+    privileged: true
+    cap_add:
+      - NET_ADMIN
+      - SYS_MODULE
+    ports:
+      - "80:80"
+    networks:
+      agent-network:
+        ipv4_address: 169.254.169.254
+      wireserver-network:
+        ipv4_address: 168.63.129.16
+    environment:
+      - RUNNING_IN_DOCKER=true
+    healthcheck:
+      test: ["CMD", "curl", "-f", "http://localhost:80/metadata/instance?api-version=2021-02-01"]
+      interval: 30s
+      timeout: 10s
+      retries: 3
+      start_period: 40s
+    logging:
+      driver: "json-file"
+      options:
+        max-size: "10m"
+        max-file: "3"
+
+networks:
+  agent-network:
+    driver: bridge
+    name: imds-network
+    ipam:
+      config:
+        - subnet: 169.254.0.0/16
+          gateway: 169.254.0.1
+  
+  wireserver-network:
+    driver: bridge
+    name: wireserver-network
+    ipam:
+      config:
+        - subnet: 168.63.0.0/16
+          gateway: 168.63.0.1
